+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 CHANGELOG
 =========
 
-1.8, 2023-07-10
+1.8, 2023-07-11
 ---------------
 - add fullscreen view to mouse heatmaps
+- add applications data to statistics pages (issue #23)
 - add option to blacklist or whitelist applications (issue #23, #25)
 - add option to configure regions of interest and disinterest for mouse events (issue #25)
 - store information on the active application in database together with input events